+++ partner/src/epilot/models/operations/invitepartner.py
@@ -22,9 +22,12 @@ class InvitePartnerRequest:
 @dataclasses.dataclass
 class InvitePartnerResponse:
     content_type: str = dataclasses.field()
+    r"""HTTP response content type for this operation"""
     status_code: int = dataclasses.field()
+    r"""HTTP response status code for this operation"""
     partner: Optional[shared_partner.Partner] = dataclasses.field(default=None)
     r"""Invited successfully"""
     raw_response: Optional[requests_http.Response] = dataclasses.field(default=None)
+    r"""Raw HTTP response; suitable for custom response parsing"""