+++ partner/src/epilot/models/operations/rejectpartner.py
@@ -20,9 +20,12 @@ class RejectPartnerRequest:
 @dataclasses.dataclass
 class RejectPartnerResponse:
     content_type: str = dataclasses.field()
+    r"""HTTP response content type for this operation"""
     status_code: int = dataclasses.field()
+    r"""HTTP response status code for this operation"""
     partner: Optional[shared_partner.Partner] = dataclasses.field(default=None)
     r"""Invited successfully"""
     raw_response: Optional[requests_http.Response] = dataclasses.field(default=None)
+    r"""Raw HTTP response; suitable for custom response parsing"""