+++ partner/src/epilot/models/operations/resendpartnerinvitation.py
@@ -39,9 +39,12 @@ class ResendPartnerInvitationRequest:
 @dataclasses.dataclass
 class ResendPartnerInvitationResponse:
     content_type: str = dataclasses.field()
+    r"""HTTP response content type for this operation"""
     status_code: int = dataclasses.field()
+    r"""HTTP response status code for this operation"""
     partner: Optional[shared_partner.Partner] = dataclasses.field(default=None)
     r"""Partner Invitation sent successfully"""
     raw_response: Optional[requests_http.Response] = dataclasses.field(default=None)
+    r"""Raw HTTP response; suitable for custom response parsing"""