+++ partner/src/epilot/models/operations/searchassignable.py
@@ -48,8 +48,11 @@ class SearchAssignable200ApplicationJSON:
 @dataclasses.dataclass
 class SearchAssignableResponse:
     content_type: str = dataclasses.field()
+    r"""HTTP response content type for this operation"""
     status_code: int = dataclasses.field()
+    r"""HTTP response status code for this operation"""
     raw_response: Optional[requests_http.Response] = dataclasses.field(default=None)
+    r"""Raw HTTP response; suitable for custom response parsing"""
     search_assignable_200_application_json_object: Optional[SearchAssignable200ApplicationJSON] = dataclasses.field(default=None)
     r"""List of assignable results"""