+++ partner/src/epilot/models/operations/searchgeolocationfortext.py
@@ -11,9 +11,12 @@ from typing import Optional
 @dataclasses.dataclass
 class SearchGeolocationForTextResponse:
     content_type: str = dataclasses.field()
+    r"""HTTP response content type for this operation"""
     status_code: int = dataclasses.field()
+    r"""HTTP response status code for this operation"""
     geolocation: Optional[shared_geolocation.Geolocation] = dataclasses.field(default=None)
     r"""Geo-location converted from text"""
     raw_response: Optional[requests_http.Response] = dataclasses.field(default=None)
+    r"""Raw HTTP response; suitable for custom response parsing"""